requirements.txt
# Core dependencies
google-genai~=1.0.0
google-adk~=0.0.1
openai~=1.75.0
supabase>=2.0.0
litellm~=1.66.3
httpx>=0.24.0,<0.25.0

# Web and UI dependencies
dash~=3.0.3
dash-mantine-components~=1.1.1
dash-core-components~=2.0.0
Flask>=1.0.4,<3.1

# Audio processing dependencies
PyAudio~=0.2.14
websockets>=13.0,<15.0dev
wave~=0.0.2

# Data processing and utilities
numpy>=1.21,<2
python-dotenv~=1.1.0
matplotlib~=3.8.3
requests~=2.32.3
protobuf~=6.30.2
scipy~=1.12.0
pandas~=2.2.0
nltk~=3.8.1

# Image processing
Pillow~=11.2.1
mss~=10.0.0

# NLP dependencies
spacy~=3.8.5

# Database dependencies
psycopg2~=2.9.10
